Goodbye enum :disappointed:
Removed enum and added separate function for one and many functions
